+++ app/com/crackers/model/Category.java
@@ -7,12 +7,11 @@ import lombok.Data;
 
 @Data
 @NodeEntity(label = "Category")
-public class Category implements java.io.Serializable
-{
+public class Category implements java.io.Serializable {
 
 	private static final long	serialVersionUID	= 1L;
 	@GraphId
-	private Integer				id;
+	private Long				id;
 	private String				categoryName;
 	private Short				isDeleted;
 	private Integer				sortOrder;
