+++ app/com/crackers/model/FunctionalAccess.java
@@ -7,12 +7,11 @@ import lombok.Data;
 
 @Data
 @NodeEntity(label = "FunctionalAccess")
-public class FunctionalAccess implements java.io.Serializable
-{
+public class FunctionalAccess implements java.io.Serializable {
 
 	private static final long	serialVersionUID	= 1L;
 	@GraphId
-	private Integer				id;
+	private Long				id;
 	private String				functionAccess;
 	private Integer				createdBy;
 	private Long				createdOn;
