+++ app/com/crackers/model/FunctionalResource.java
@@ -7,12 +7,11 @@ import lombok.Data;
 
 @Data
 @NodeEntity(label = "FunctionalResource")
-public class FunctionalResource implements java.io.Serializable
-{
+public class FunctionalResource implements java.io.Serializable {
 
 	private static final long	serialVersionUID	= 1L;
 	@GraphId
-	private Integer				id;
+	private Long				id;
 	private String				functions;
 	private Integer				createdBy;
 	private Long				createdOn;
