+++ app/com/crackers/model/ImageColorCode.java
@@ -7,12 +7,11 @@ import lombok.Data;
 
 @Data
 @NodeEntity(label = "ImageColorCode")
-public class ImageColorCode implements java.io.Serializable
-{
+public class ImageColorCode implements java.io.Serializable {
 
 	private static final long	serialVersionUID	= 1L;
 	@GraphId
-	private Integer				id;
+	private Long				id;
 	private String				imageColorCodeValue;
 	private Short				isDeleted;
 	private Integer				createdBy;
